requirements.txt
paramiko~=3.3.1
python-dotenv~=1.0.0
pynetbox~=7.2.0
netaddr~=0.9.0